+++ gateway/src/apicast/policy/ip_check/client_ip.lua
@@ -13,6 +13,11 @@ end
 
 local function ip_from_x_forwarded_for_header()
   local forwarded_for = ngx.req.get_headers()['X-Forwarded-For']
+  -- If the header is duplicated, get_heders() method  returns all the headers
+  -- in a table instead of the string, so we only use the first one.
+  if (type(forwarded_for) == "table") then
+    forwarded_for = forwarded_for[1]
+  end
 
   if not forwarded_for or forwarded_for == "" then
     return nil